refactor: pr 수정사항 적용 완료
