+++ src/app/(routes)/page.tsx
@@ -1,5 +1,6 @@
+
 const LandingPage = () => {
-  return <div className='text-4xl-bold md:text-2xl-bold'>Hello World</div>;
+  return  <div className='text-4xl-bold md:text-2xl-bold'>Hello World</div>
 };
 
 export default LandingPage;
