+++ src/app/loading.tsx
@@ -1,12 +1,8 @@
-import Spinner from '@/shared/components/ui/spinner/Spinner';
+import Spinner from '@/shared/components/ui/Spinner';
 
 const Loading = () => {
   return (
     <div className='flex flex-col items-center justify-center h-screen'>
-      <p className='text-xl-bold'>로딩 UI test 3초...</p>
-      <br />
-      <br />
-      <br />
       <Spinner />
     </div>
   );
